pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bento"
version = "0.1.0"
description = "Popup control centre for KDE Plasma"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [{name = "HabiRabbu"}]
keywords = ["kde", "plasma", "tray", "popup", "wayland"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: X11 Applications :: Qt",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Desktop Environment :: K Desktop Environment (KDE)",
]
dependencies = [
    "PyQt6>=6.5",
    "dbus-python>=1.3",
    "python-dotenv>=1.0",
    "pyte>=0.8",
]

[project.urls]
Homepage = "https://github.com/HabiRabbu/bento"
Repository = "https://github.com/HabiRabbu/bento"
Issues = "https://github.com/HabiRabbu/bento/issues"

[project.scripts]
bento = "bento_app.__main__:main"

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-qt>=4.0", "mypy>=1.5", "ruff>=0.1"]

[tool.setuptools.packages.find]
include = ["bento_app*"]

[tool.setuptools.package-data]
bento_app = [
    "assets/*.svg",
    "assets/*.png",
    "builtin_blocks/*/manifest.json",
    "builtin_blocks/*/block.py",
    "builtin_blocks/*/icon.svg",
    "py.typed",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
no_implicit_optional = true

[tool.ruff]
target-version = "py311"
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
